Программа про Машу

+++ Examol004_Div/Program.cs
@@ -1,3 +1,3 @@
-﻿int numberA = 10;
-int numberB = 5;
+﻿double numberA = 10;
+double numberB = 5;
 Console.Writeline(numberA / numberB);
